please send help
for some reason has(xpos) = true but doesn't return a value other than 0

hate those stupid nodes
